Upgrade swarm-admin to 0.4.0, removed Docker:dind
In swarm-admin 0.4.0, the pipeline's containers automatically join
app-swarm-ui's network when they are started. This change requires
that the swarm-admin and the pipelines run on the same Docker daemon.

Also, by default the network of app-swarm-ui must be overlay in order to
be shared across the swarm. But for development purpose, this is
overridden in the override file to use a bridge network and to work on a
regular Docker daemon.